Move surveys logic into their own struct
